add table/api benchmarks
