+++ src/lib.rs
@@ -5,3 +5,6 @@ pub mod discovery;
 pub mod manager;
 pub mod parquet;
 pub mod table;
+
+#[cfg(feature = "testutil")]
+pub mod testutil;
